+++ src/constants.ts
@@ -2,6 +2,7 @@ export const PLUGIN_NAME = '@nasti-toolchain/plugin-vue-lynx'
 export const PLUGIN_VERSION = '0.1.0'
 export const RSPEEDY_DRIVER = 'rspeedy'
 export const CLIENT_BRIDGE_DRIVER = `${PLUGIN_NAME}:client-bridge`
+export const NATIVE_SERVE_DRIVER = `${PLUGIN_NAME}:native-serve`
 export const NASTI_BACKGROUND_ENVIRONMENT = 'lynx-background'
 export const NASTI_MAIN_THREAD_ENVIRONMENT = 'lynx-main-thread'